Move cbor package documentation to doc.go
